llm/requirements.txt
openai
fastapi
uvicorn[all]
aiohttp
loguru
python-dotenv